use ModalContent to populate modal
Using some structural patterns, the different possible contents of
the modal is encapsulated into their own components and instantiated
depending on the application's `mode`.

The main benefit of this system is reuse of the single Modal
and standardization of the props passed to it. It is also easier to
make a standard dialoge yes/no modal that may optionally have
additional fields as part of the Yes.

On the other hand, some complexity is introduced. A store like Redux
would have made this setup significantly simpler.